Reduce inter-module imports
